use double val for sums

+++ src/JoinOrder.cpp
@@ -18,7 +18,7 @@ JoinOrder::JoinOrder(uint32_t tables) :
         orderedTables(0),
         tableOrder(new uint32_t[tables]),
         stats(new const MultipleColumnStats*[tables]),
-        rowSum(new uint64_t[tables]) {
+        rowSum(new double[tables]) {
 }
 
 JoinOrder::JoinOrder(uint32_t tables,
@@ -33,11 +33,11 @@ JoinOrder::JoinOrder(const JoinOrder& toCopy) :
         orderedTables(toCopy.orderedTables),
         tableOrder(new uint32_t[toCopy.arraySize]),
         stats(new const MultipleColumnStats*[toCopy.arraySize]),
-        rowSum(new uint64_t[toCopy.arraySize]) {
+        rowSum(new double[toCopy.arraySize]) {
     memcpy(tableOrder,
            toCopy.tableOrder,
            toCopy.orderedTables * sizeof(uint32_t));
-    memcpy(rowSum, toCopy.rowSum, toCopy.orderedTables * sizeof(uint64_t));
+    memcpy(rowSum, toCopy.rowSum, toCopy.orderedTables * sizeof(double));
     for (uint32_t i = 0; i < toCopy.orderedTables; ++i) {
         stats[i] = new MultipleColumnStats(*(toCopy.stats[i]));
     }
@@ -68,7 +68,7 @@ JoinOrder& JoinOrder::operator=(const JoinOrder& toCopy) {
     memcpy(tableOrder,
            toCopy.tableOrder,
            toCopy.orderedTables * sizeof(uint32_t));
-    memcpy(rowSum, toCopy.rowSum, toCopy.orderedTables * sizeof(uint64_t));
+    memcpy(rowSum, toCopy.rowSum, toCopy.orderedTables * sizeof(double));
     for (uint32_t i = 0; i < toCopy.orderedTables; ++i) {
         stats[i] = new MultipleColumnStats(*(toCopy.stats[i]));
     }
@@ -186,7 +186,7 @@ const MultipleColumnStats* const * JoinOrder::getStats() const {
     return stats;
 }
 
-const uint64_t* JoinOrder::getRowSum() const {
+const double* JoinOrder::getRowSum() const {
     return rowSum;
 }
